// constants/scroll-transform-exploration/element-id.enum.ts
enum ElementID {
    // hello_world
    ANIMATION_CONTAINER = 'animation_container',
    HELLO_WORLD = 'hello_world',
    // hello_world_level_0
    HELLO_WORLD_LEVEL_0 = 'hello_world_level_0',
    // hello_world_level_1
    HELLO_WORLD_LEVEL_1 = 'hello_world_level_1',
    HOME_SCREEN = 'home_screen',
    SCROLL_CONT = 'scroll_cont',
    // hello_world_level_2
    HELLO_WORLD_LEVEL_2 = 'hello_world_level_2',
    SHEET = 'sheet',
    // hello_world_level_3
    HELLO_WORLD_LEVEL_3 = 'hello_world_level_3',
    HELLO_WORLD_CONT = 'hello_world_cont',
    INTROS = 'intros',
    IM_NATE_SMITH_CONT = 'im_nate_smith_cont',
    A_CONT = 'a_cont',
    LOS_ANGELES_BASED_CONT = 'los_angeles_based_cont',
    DESIGNER_CONT = 'designer_cont',
    METICULOUS_DESIGNER_CONT = 'meticulous_designer_cont',
    FLUID_INTERFACE_CONT = 'fluid_interface_cont',
    CURRENT_JOB_CONT = 'current_job_cont',
    SCROLL_FOR_WORK_CONT = 'scroll_for_work_cont',
    HOME_BAR = 'home_bar',

    // romento
    ROMENTO = 'romento',
    ROMENTO_LEVEL_0 = 'romento_level_0',
    ROMENTO_LEVEL_1 = 'romento_level_1',
    ROMENTO_VIDEO = 'romento_video',
    ROMENTO_LEVEL_2 = 'romento_level_2',
    ROMENTO_LEVEL_3 = 'romento_level_3',

    // hulu tv
    HULU_TV = 'hulu_tv',
    HULU_TV_LEVEL_0 = 'hulu_tv_level_0',
    HULU_TV_STAND = 'hulu_tv_stand',
    HULU_TV_LEVEL_1 = 'hulu_tv_level_1',
    HULU_TV_VIDEO_CONT = 'hulu_tv_video_cont',
    HULU_TV_LEVEL_2 = 'hulu_tv_level_2',
    HULU_TV_LEVEL_3 = 'hulu_tv_level_3',
    HULU_TV_LEVEL_4 = 'hulu_tv_level_4',

    // laptop-1
    LAPTOP_1 = 'laptop_1',
    LAPTOP_1_LEVEL_0 = 'laptop_1_level_0',
    LAPTOP_1_BOTTOM_CONT = 'laptop_1_bottom_cont',
    LAPTOP_1_BOTTOM = 'laptop_1_bottom',
    LAPTOP_1_LEVEL_1 = 'laptop_1_level_1',
    LAPTOP_1_VIDEO_CONT = 'laptop_1_video_cont',
    LAPTOP_1_LEVEL_2 = 'laptop_1_level_2',
    LAPTOP_1_LEVEL_3 = 'laptop_1_level_3',

    // laptop 2
    LAPTOP_2 = 'laptop_2',
    LAPTOP_WRAPPER = 'laptop_wrapper',
    LAPTOP_2_LEVEL_0 = 'laptop_2_level_0',
    LAPTOP_2_TOP = 'laptop_2_top',
    LAPTOP_2_BOTTOM_CONT = 'laptop_2_bottom_cont',
    LAPTOP_2_BOTTOM = 'laptop_2_bottom',
    LAPTOP_2_LEVEL_1 = 'laptop_2_level_1',
    LAPTOP_2_LEVEL_2 = 'laptop_2_level_2',
    LAPTOP_2_LEVEL_3 = 'laptop_2_level_3',

    // tablet
    TABLET = 'tablet',
    TABLET_LEVEL_0 = 'tablet_level_0',
    TABLET_LEVEL_1 = 'tablet_level_1',
    TABLET_LEVEL_2 = 'tablet_level_2',
    TABLET_LEVEL_3 = 'tablet_level_3',
    CONCLUSION_1 = 'conclusion_1',
    CONCLUSION_2 = 'conclusion_2',
}

// eslint-disable-next-line no-shadow
enum ScrollElementIDEnum {
    SHEET_IN = 'sheet_in',
    HOME_BAR_COLOR = 'home_bar_color',
    TEXT_SCROLL_FADE_OUT = 'text_scroll_fade_out',
    HOME_SCREEN_SCALE = 'home_screen_scale',
    HELLO_WORLD_IN = 'hello_world_in',
    IM_NATE_SMITH_IN = 'im_nate_smith_in',
    LOS_ANGELES_BASED_IN = 'los_angeles_based_in',
    METICULOUS_VISUAL_IN = 'meticulous_visual_in',
    FLUID_INTERFACE_IN = 'fluid_interface_in',
    A_DESIGNER_IN = 'a_designer_in',
    CURRENT_JOB_IN = 'current_job_in',
    SCROLL_FOR_WORK_IN = 'scroll_for_work_in',
    HELLO_OUT = 'hello_out',
    HELLO_FADE_OUT = 'hello_fade_out',
    ROMENTO_IN = 'romento_in',
    ROMENTO_FADE_IN = 'romento_fade_in',
    PHONE_OUT = 'phone_out',
    PHONE_ROTATE_X = 'phone_rotate_x',
    PHONE_ROTATE_Z = 'phone_rotate_z',
    PHONE_FADE_OUT = 'phone_fade_out',
    TV_IN = 'tv_in',
    TV_FADE_IN = 'tv_fade_in',
    TV_ROTATE_Z = 'tv_rotate_z',
    TV_ROTATE_X = 'tv_rotate_x',
    TV_LEGS_OUT = 'tv_legs_out',
    TV_OUT = 'tv_out',
    TV_FADE_OUT = 'tv_fade_out',
    LAPTOP_1_IN = 'laptop_1_in',
    LAPTOP_1_FADE_IN = 'laptop_1_fade_in',
    BOTTOM_POS = 'bottom_pos',
    BOTTOM_POS_0 = 'bottom_pos_0',
    BOTTOM_POS_1 = 'bottom_pos_1',
    BOTTOM_POS_2 = 'bottom_pos_2',
    LAPTOP_1_BOTTOM_FADE_IN = 'laptop_1_bottom_fade_in',
    LAPTOP_1_OUT = 'laptop_1_out',
    BOTTOM_CLOSE_1 = 'bottom_close_1',
    BOTTOM_TOP_FADE_OUT = 'bottom_top_fade_out',
    BOTTOM_CLOSE_2 = 'bottom_close_2',
    LAPTOP_1_FADE_OUT = 'laptop_1_fade_out',
    LAPTOP_2_IN = 'laptop_2_in',
    LAPTOP_2_FADE_IN = 'laptop_2_fade_in',
    LAPTOP_2_BOTTOM_FADE_IN = 'laptop_2_bottom_fade_in',
    LAPTOP_2_OUT = 'laptop_2_out',
    LAPTOP_2_OPEN_FLAT = 'laptop_2_open_flat',
    LAPTOP_2_FOLD_BACKWARDS = 'laptop_2_fold_backwards',
    LAPTOP_2_FADE_OUT = 'laptop_2_fade_out',
    TABLET_IN = 'tablet_in',
    TABLET_FADE_IN = 'tablet_fade_in',
    TABLET_OUT = 'tablet_out',
    TABLET_FADE_OUT = 'tablet_fade_out',
    CONCLUSION_1_IN = 'conclusion_1_in',
    CONCLUSION_2_IN = 'conclusion_2_in',
}